Safari-609.1.11.1/LayoutTests/crypto/workers/subtle/ec-generate-export-jwk-key-expected.txt

[Worker] Test exporting an ECDH key pair with JWK format in workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/ec-generate-export-jwk-key.js
[Worker] Generating a key pair...
[Worker] Exporting the public key...
PASS [Worker] publicKey.kty is 'EC'
PASS [Worker] publicKey.crv is 'P-384'
PASS [Worker] Base64URL.parse(publicKey.x).byteLength is 48
PASS [Worker] Base64URL.parse(publicKey.y).byteLength is 48
PASS [Worker] publicKey.d is undefined.
PASS [Worker] publicKey.key_ops is [ ]
PASS [Worker] publicKey.ext is true
[Worker] Exporting the private key...
PASS [Worker] privateKey.kty is 'EC'
PASS [Worker] privateKey.crv is 'P-384'
PASS [Worker] Base64URL.parse(privateKey.x).byteLength is 48
PASS [Worker] Base64URL.parse(privateKey.y).byteLength is 48
PASS [Worker] Base64URL.parse(privateKey.d).byteLength is 48
PASS [Worker] privateKey.key_ops is ['deriveBits', 'deriveKey']
PASS [Worker] privateKey.ext is true
PASS successfullyParsed is true

TEST COMPLETE
